+++ addons/base_setup/models/res_config_settings.py
@@ -22,7 +22,7 @@ class ResConfigSettings(models.TransientModel):
         string='Allow the users to synchronize their calendar  with Google Calendar')
     module_microsoft_calendar = fields.Boolean(
         string='Allow the users to synchronize their calendar with Outlook Calendar')
-    module_mail_client_extension = fields.Boolean(
+    module_mail_plugin = fields.Boolean(
         string='Allow integration with the mail plugins'
     )
     module_google_drive = fields.Boolean("Attach Google documents to any record")
