+++ addons/base_setup/views/res_config_settings_views.xml
@@ -210,12 +210,12 @@
 
                         <h2>Integrations</h2>
                         <div class="row mt16 o_settings_container" name="integration">
-                            <div class="col-12 col-lg-6 o_setting_box" id="mail_client_extension_setting">
+                            <div class="col-12 col-lg-6 o_setting_box" id="mail_pluggin_setting">
                                 <div class="o_setting_left_pane">
-                                    <field name="module_mail_client_extension" />
+                                    <field name="module_mail_plugin" />
                                 </div>
                                 <div class="o_setting_right_pane">
-                                    <label string="Mail Client Extension" for="module_mail_client_extension"/>
+                                    <label string="Mail Plugin" for="module_mail_plugin"/>
                                     <a href="https://www.odoo.com/documentation/user/crm/optimize/outlook_extension.html" title="Documentation" class="o_doc_link" target="_blank"></a>
                                     <div class="text-muted">
                                         Integrate with mail client plugins
