+++ addons/crm_mail_plugin/__init__.py
