+++ addons/crm_mail_plugin/controllers/__init__.py
