+++ addons/crm_mail_plugin/controllers/mail_plugin.py
@@ -6,7 +6,7 @@ import logging
 from odoo.http import request
 from odoo.tools.misc import formatLang
 
-from odoo.addons.mail_client_extension.controllers import mail_plugin
+from odoo.addons.mail_plugin.controllers import mail_plugin
 
 _logger = logging.getLogger(__name__)
