+++ addons/crm_mail_plugin/models/__init__.py
