+++ addons/crm_mail_plugin/views/crm_lead_views.xml
