+++ addons/crm_mail_plugin/views/crm_mail_plugin_lead.xml
