+++ addons/mail_plugin/i18n/mail_client_extension.pot
