+++ addons/mail_plugin/models/__init__.py
