+++ addons/mail_plugin/views/mail_plugin_login.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
-    <template id="mail_client_extension.app_auth" name="Accept app">
+    <template id="mail_plugin.app_auth" name="Accept app">
         <t t-call="web.login_layout">
             <t t-set="disable_database_manager" t-value="1"/>
-            <form class="oe_login_form" role="form" method="post" action="/mail_client_extension/auth/confirm">
+            <form class="oe_login_form" role="form" method="post" action="/mail_plugin/auth/confirm">
                 <input type="hidden" name="csrf_token" t-att-value="request.csrf_token()"/>
                 <input type="hidden" name="redirect" t-att-value="redirect"/>
                 <input type="hidden" name="scope" t-att-value="scope"/>
